refactor dag compiler code to separate file
